+++ moveit_ros/visualization/robot_state_rviz_plugin/src/robot_state_display.cpp
@@ -457,7 +457,7 @@ void RobotStateDisplay::onDisable()
   Display::onDisable();
 }
 
-void RobotStateDisplay::update(double wall_dt, double ros_dt)
+void RobotStateDisplay::update(float wall_dt, float ros_dt)
 {
   Display::update(wall_dt, ros_dt);
   calculateOffsetPosition();
